requirements.txt
Django==2.2
djangorestframework-csv==2.1.0
mimesis==3.3.0
psycopg2-binary==2.8.4
